@@ -1,6 +1,8 @@
-export interface User{
-    uid: string;
-    email: string;
-    password: string;
-    name: string;
+export interface User {
+  uid: string;
+  name?: string;
+  email: string;
+  password?: string;
+  role?: 'supervisor' | 'encargado' | 'pollero';
+  assignedShed?: string | null;
 }
